Moved code from PoC to Folio space
